Complete refactor of module into more logical structure
