+++ neon_enclosure/components/display_manager.py
