+++ neon_enclosure/components/hardware/MycroftFan/fan_sj201r4.py
