+++ neon_enclosure/components/hardware/MycroftLed/default_palette.py
