+++ neon_enclosure/components/hardware/MycroftLed/led_neo_pixel.py
