+++ neon_enclosure/components/hardware/MycroftLed/led_sj201r4.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import os
-from neon_enclosure.enclosure.hardware.MycroftLed.MycroftLed import MycroftLed
+from neon_enclosure.components.hardware.MycroftLed.MycroftLed import MycroftLed
 from neon_utils import LOG
