+++ neon_enclosure/components/hardware/MycroftSwitch/MycroftSwitch.py
