+++ neon_enclosure/components/hardware/MycroftSwitch/__init__.py
