+++ neon_enclosure/components/hardware/MycroftSwitch/switch_gpio.py
